refactor(platform): move getPlatformType function to Platform folder
