--- composeApp/src/commonMain/kotlin/org/collcal/collcal/Platform.kt
@@ -1,9 +1,0 @@
-package org.collcal.collcal
-
-enum class PlatformType {
-    ANDROID,
-    WEB,
-    IOS
-}
-
-expect fun getPlatformType(): PlatformType
