+++ composeApp/src/commonMain/kotlin/org/collcal/collcal/platform/PlatformType.kt
@@ -1,0 +1,8 @@
+package org.collcal.collcal.platform
+
+enum class PlatformType {
+    ANDROID,
+    WEB,
+    IOS
+}
+
