+++ composeApp/src/commonMain/kotlin/org/collcal/collcal/platform/getPlatformType.kt
@@ -1,0 +1,3 @@
+package org.collcal.collcal.platform
+
+expect fun getPlatformType(): PlatformType
