+++ composeApp/src/commonMain/kotlin/org/collcal/collcal/presentation/onboarding/OnBoardingScreen.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.collcal.collcal.PlatformType
-import org.collcal.collcal.getPlatformType
+import org.collcal.collcal.platform.PlatformType
+import org.collcal.collcal.platform.getPlatformType
 import org.collcal.collcal.navigation.Navigator
 import org.collcal.collcal.navigation.Screen
 import org.collcal.collcal.presentation.ui.theme.Strings
