--- composeApp/src/iosMain/kotlin/org/collcal/collcal/platform/getScreenWidthDp.ios.kt
@@ -1,8 +1,0 @@
-package org.collcal.collcal.platform
-
-import androidx.compose.runtime.Composable
-
-@Composable
-actual fun getScreenWidthDp(): Int {
-    return 0
-}
